Added the functionality to upload to controller a file,

+++ Server/src/main/java/com/trading212/judge/web/controller/DocumentController.java
@@ -5,12 +5,15 @@ import com.trading212.judge.model.dto.DocumentDTO;
 import com.trading212.judge.model.dto.DocumentSimpleDTO;
 import com.trading212.judge.model.dto.TaskSimpleDTO;
 import com.trading212.judge.service.task.DocumentService;
+import com.trading212.judge.util.path.ResourcePathUtil;
 import com.trading212.judge.web.exception.ServiceUnavailableException;
+import com.trading212.judge.web.exception.UnexpectedFailureException;
 import com.trading212.judge.web.exception.task.DocumentDataException;
 import com.trading212.judge.service.task.TaskService;
 import com.trading212.judge.web.exception.task.DocumentExistException;
+import com.trading212.judge.web.exception.task.DocumentNotFoundException;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
-import org.springframework.boot.autoconfigure.web.servlet.error.BasicErrorController;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
@@ -18,8 +21,6 @@ import org.springframework.web.bind.annotation.*;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.URI;
-import java.time.Instant;
 import java.util.Optional;
 import java.util.Set;
 
@@ -29,10 +30,12 @@ public class DocumentController {
 
     private final TaskService taskService;
     private final DocumentService documentService;
+    private final ResourcePathUtil resourcePathUtil;
 
-    public DocumentController(TaskService taskService, DocumentService documentService) {
+    public DocumentController(TaskService taskService, DocumentService documentService, ResourcePathUtil resourcePathUtil) {
         this.taskService = taskService;
         this.documentService = documentService;
+        this.resourcePathUtil = resourcePathUtil;
     }
 
     @GetMapping
@@ -44,7 +47,7 @@ public class DocumentController {
 
     @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<DocumentDTO> create(@ModelAttribute @Valid DocumentBindingModel documentBindingModel,
-                                              BindingResult bindingResult) {
+                                              BindingResult bindingResult, HttpServletRequest httpServletRequest) {
 
         if (bindingResult.hasErrors()) {
             throw new DocumentDataException("Invalid data!");
@@ -60,10 +63,9 @@ public class DocumentController {
 
         try {
             tempDocFile = File.createTempFile("document", "docx");
-            documentBindingModel.file().transferTo(tempDocFile);
+            documentBindingModel.document().transferTo(tempDocFile);
         } catch (IOException e) {
-            e.printStackTrace();
-            return ResponseEntity.internalServerError().build();
+            throw new UnexpectedFailureException("Temp file creation exception");
         }
 
         Optional<DocumentDTO> document = documentService.create(documentBindingModel.name(), documentBindingModel.isTest(),
@@ -75,18 +77,36 @@ public class DocumentController {
             throw new ServiceUnavailableException("Service is shut down temporarily.");
         }
 
-        return ResponseEntity.created(URI.create("someurl")).body(document.get());
+        return ResponseEntity.created(resourcePathUtil.buildResourcePath(httpServletRequest, document.get().id()))
+                .body(document.get());
     }
 
     @DeleteMapping(path = Routes.BY_ID)
-    public ResponseEntity<?> delete(@PathVariable Integer id) {
+    public ResponseEntity<?> deleteById(@PathVariable Integer id) {
+        boolean exist = documentService.isExist(id);
+
+        if (!exist) {
+            throw new DocumentNotFoundException("Document does not exist");
+        }
+
         boolean isDeleted = documentService.delete(id);
 
         if (!isDeleted) {
-            throw new DocumentDataException("Data invalid");
+            throw new ServiceUnavailableException("Document cannot be deleted!");
         }
 
-        return ResponseEntity.ok().build();
+        return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping(path = Routes.BY_ID)
+    public ResponseEntity<?> findById(@PathVariable Integer id) {
+        Optional<DocumentDTO> documentDTO = documentService.findByID(id);
+
+        if (documentDTO.isEmpty()) {
+            throw new DocumentNotFoundException("Document could not be found!");
+        }
+
+        return ResponseEntity.ok(documentDTO);
     }
 
     @GetMapping(path = Routes.FIND_ALL_TASKS_BY_DOCUMENT_ID)
